Renaming the structure.py to grid.py
