+++ test/Algorithm/algorithm_astar_test.py
@@ -1,6 +1,6 @@
 import unittest
 from modules.CAN_Navigation.module.Algorithms import AStar
-from modules.CAN_Navigation.module.Structure import grid_factory
+from modules.CAN_Navigation.module.Grid import grid_factory
 
 
 class TestAlgorithmAstar(unittest.TestCase):
