+++ test/Grid/grid_inmemory_test.py
@@ -1,5 +1,5 @@
 import unittest
-from modules.CAN_Navigation.module.Structure import GridInMemory
+from modules.CAN_Navigation.module.Grid import GridInMemory
 
 
 class TestGridInMemory(unittest.TestCase):
